Create Gridsearchcv run for best parameters
For this part, a gridsearchcv was set up to search for the best parameters regarding batch size, epochs, learning rate and optimizer. The test looked through every combination and a log file was set up to return the results for each interaction, ranking them by test value score (which was set up using R² as a parameter).