Create Ann test with best parameters
By running the model several times, the ann was able to be tuned starting from the parameters defined by gridsearchcv, slowly increasing epochs and testing layer sizes. Finally, test size data was changed to 0,15 and the first layers were increased, returning a final R² value of 0,97 and MSLE value of 0,11. A log file was created for all real and predicted values for Qlim.